refactor: validation
